feature/post add error code to delete API

+++ src/main/java/com/example/debriserver/core/Post/PostService.java
@@ -23,6 +23,9 @@ public class PostService {
         this.postProvider = postProvider;
     }
 
+    /**
+     * 게시물 생성
+     */
     public PostPostsRes createPosts(PostPostsReq postPostsReq) throws BasicException {
 
         try{
@@ -41,6 +44,9 @@ public class PostService {
         }
     }
 
+    /**
+     * 게시물 수정
+     */
     public void modifyPost(int userIdx, int postIdx, PatchPostsReq patchPostsReq) throws BasicException {
         if (postProvider.checkUserExist(userIdx) == 0) {
             throw new BasicException(USERS_EMPTY_USER_ID);
@@ -62,9 +68,16 @@ public class PostService {
         }
     }
 
+    /**
+     * 게시물 삭제
+     */
     public void deletePost(int postIdx) throws BasicException {
 
         try{
+
+            if (postProvider.checkPostExist(postIdx) == 0) {
+                throw new BasicException(POSTS_EMPTY_POST_ID);
+            }
 
             int result = postDao.deletePost(postIdx);
             if (result == 0) {
